+++ apps/webapp/src/app/api/app/profile/connect/google/callback/route.ts
@@ -1,8 +1,10 @@
+import { api } from '@workspace/backend/convex/_generated/api';
+import { fetchAction } from 'convex/nextjs';
 import { type NextRequest, NextResponse } from 'next/server';
 
 /**
- * Google OAuth callback for profile connect - proxies to Convex HTTP endpoint
- * This allows the callback to come from the app's domain instead of Convex's domain
+ * Google OAuth callback for profile connect - calls Convex action directly using type-safe fetchAction
+ * This processes the OAuth callback and handles the account linking flow
  */
 export async function GET(request: NextRequest) {
   const url = new URL(request.url);
@@ -21,42 +23,26 @@ export async function GET(request: NextRequest) {
   }
 
   try {
-    // Get the Convex backend URL and convert to site URL for HTTP endpoints
-    const convexUrl = process.env.NEXT_PUBLIC_CONVEX_URL;
-    if (!convexUrl) {
-      throw new Error('NEXT_PUBLIC_CONVEX_URL environment variable is not set');
+    // Call the Convex action using type-safe fetchAction
+    const result = await fetchAction(api.auth.google.handleGoogleConnectCallback, {
+      code,
+      state,
+    });
+
+    if (result.success) {
+      // Return HTML/JS to close the window
+      return new NextResponse(
+        '<html><body><script>window.close();</script>Account connected successfully. You may close this window.</body></html>',
+        { status: 200, headers: { 'Content-Type': 'text/html' } }
+      );
     }
-
-    // Convert .cloud URL to .site URL for HTTP endpoints
-    const convexSiteUrl = convexUrl.replace('.cloud', '.site');
-
-    // Construct the Convex HTTP endpoint URL
-    const convexHttpUrl = `${convexSiteUrl}/app/profile/connect/google/callback`;
-    const convexCallbackUrl = new URL(convexHttpUrl);
-    convexCallbackUrl.searchParams.set('code', code);
-    convexCallbackUrl.searchParams.set('state', state);
-
-    // Forward the request to Convex
-    const response = await fetch(convexCallbackUrl.toString(), {
-      method: 'GET',
-      headers: {
-        'User-Agent': request.headers.get('User-Agent') || 'NextJS-Proxy/1.0',
-      },
-    });
-
-    // Get the response body and headers from Convex
-    const body = await response.text();
-    const contentType = response.headers.get('Content-Type') || 'text/html';
-
-    // Return the response from Convex
-    return new NextResponse(body, {
-      status: response.status,
-      headers: {
-        'Content-Type': contentType,
-      },
-    });
+    // Handle failure
+    return new NextResponse(
+      `<html><body>Connect failed: ${result.error || 'Unknown error'}</body></html>`,
+      { status: 500, headers: { 'Content-Type': 'text/html' } }
+    );
   } catch (error) {
-    console.error('Error proxying Google OAuth profile connect callback:', error);
+    console.error('Error calling Convex Google connect callback:', error);
     return new NextResponse(
       '<html><body>Internal server error during OAuth callback.</body></html>',
       {
