+++ apps/webapp/src/app/api/auth/google/callback/route.ts
@@ -1,8 +1,10 @@
+import { api } from '@workspace/backend/convex/_generated/api';
+import { fetchAction } from 'convex/nextjs';
 import { type NextRequest, NextResponse } from 'next/server';
 
 /**
- * Google OAuth callback for login - proxies to Convex HTTP endpoint
- * This allows the callback to come from the app's domain instead of Convex's domain
+ * Google OAuth callback for login - calls Convex action directly using type-safe fetchAction
+ * This processes the OAuth callback and handles the authentication flow
  */
 export async function GET(request: NextRequest) {
   const url = new URL(request.url);
@@ -26,61 +28,33 @@ export async function GET(request: NextRequest) {
   }
 
   try {
-    // Get the Convex backend URL and convert to site URL for HTTP endpoints
-    const convexUrl = process.env.NEXT_PUBLIC_CONVEX_URL;
-    if (!convexUrl) {
-      console.error(
-        '[Google OAuth Callback] NEXT_PUBLIC_CONVEX_URL environment variable is not set'
+    console.log(
+      '[Google OAuth Callback] Calling Convex action:',
+      'auth.google.handleGoogleLoginCallback'
+    );
+
+    // Call the Convex action using type-safe fetchAction
+    const result = await fetchAction(api.auth.google.handleGoogleLoginCallback, {
+      code,
+      state,
+    });
+
+    console.log('[Google OAuth Callback] Convex action result:', result);
+
+    if (result.success) {
+      // Return HTML/JS to close the window
+      return new NextResponse(
+        '<html><body><script>window.close();</script>Login successful. You may close this window.</body></html>',
+        { status: 200, headers: { 'Content-Type': 'text/html' } }
       );
-      throw new Error('NEXT_PUBLIC_CONVEX_URL environment variable is not set');
     }
-
-    // Convert .cloud URL to .site URL for HTTP endpoints
-    const convexSiteUrl = convexUrl.replace('.cloud', '.site');
-
-    // Construct the Convex HTTP endpoint URL
-    const convexHttpUrl = `${convexSiteUrl}/auth/google/callback`;
-    const convexCallbackUrl = new URL(convexHttpUrl);
-    convexCallbackUrl.searchParams.set('code', code);
-    convexCallbackUrl.searchParams.set('state', state);
-
-    // Debug: Log the proxied Convex callback URL
-    console.log('[Google OAuth Callback] Proxying to Convex URL:', convexCallbackUrl.toString());
-
-    // Forward the request to Convex
-    const response = await fetch(convexCallbackUrl.toString(), {
-      method: 'GET',
-      headers: {
-        'User-Agent': request.headers.get('User-Agent') || 'NextJS-Proxy/1.0',
-      },
-    });
-
-    // Debug: Log response status and headers from Convex
-    console.log('[Google OAuth Callback] Convex response status:', response.status);
-    console.log(
-      '[Google OAuth Callback] Convex response headers:',
-      Object.fromEntries(response.headers.entries())
+    // Handle failure
+    return new NextResponse(
+      `<html><body>Login failed: ${result.error || 'Unknown error'}</body></html>`,
+      { status: 500, headers: { 'Content-Type': 'text/html' } }
     );
-
-    // Get the response body and headers from Convex
-    const body = await response.text();
-    const contentType = response.headers.get('Content-Type') || 'text/html';
-
-    // Debug: Log a snippet of the response body
-    console.log(
-      '[Google OAuth Callback] Convex response body (first 200 chars):',
-      body.slice(0, 200)
-    );
-
-    // Return the response from Convex
-    return new NextResponse(body, {
-      status: response.status,
-      headers: {
-        'Content-Type': contentType,
-      },
-    });
   } catch (error) {
-    console.error('Error proxying Google OAuth callback:', error);
+    console.error('Error calling Convex Google OAuth callback:', error);
     return new NextResponse(
       '<html><body>Internal server error during OAuth callback.</body></html>',
       {
