+++ services/backend/convex/_generated/api.d.ts
@@ -22,7 +22,6 @@ import type * as cleanupTasks from "../cleanupTasks.js";
 import type * as crypto from "../crypto.js";
 import type * as discussions from "../discussions.js";
 import type * as googleAuth from "../googleAuth.js";
-import type * as http from "../http.js";
 import type * as migration from "../migration.js";
 import type * as presentations from "../presentations.js";
 import type * as serviceDesk from "../serviceDesk.js";
@@ -46,7 +45,6 @@ declare const fullApi: ApiFromModules<{
   crypto: typeof crypto;
   discussions: typeof discussions;
   googleAuth: typeof googleAuth;
-  http: typeof http;
   migration: typeof migration;
   presentations: typeof presentations;
   serviceDesk: typeof serviceDesk;
