+++ services/backend/convex/auth/google.ts
@@ -1,5 +1,8 @@
 import { v } from 'convex/values';
-import { mutation, query } from '../_generated/server';
+import type { SessionId } from 'convex-helpers/server/sessions';
+import { api } from '../_generated/api';
+import type { Id } from '../_generated/dataModel';
+import { action, mutation, query } from '../_generated/server';
 
 /**
  * Gets Google authentication configuration for client use.
@@ -124,5 +127,153 @@ export const cleanupExpiredLoginRequests = mutation({
       success: true,
       deletedCount: expiredRequests.length,
     };
+  },
+});
+
+/**
+ * Handles Google OAuth callback for login flow.
+ * Processes the OAuth code, exchanges it for a profile, logs in the user, and marks the request as completed.
+ */
+export const handleGoogleLoginCallback = action({
+  args: {
+    code: v.string(),
+    state: v.string(), // This is the loginRequestId
+  },
+  handler: async (ctx, args) => {
+    const { code, state } = args;
+
+    try {
+      // Get the login request to extract sessionId and redirectUri
+      const loginRequest = await ctx.runQuery(api.auth.google.getLoginRequest, {
+        loginRequestId: state as Id<'auth_loginRequests'>,
+      });
+      if (!loginRequest || loginRequest.provider !== 'google') {
+        throw new Error('Invalid login request');
+      }
+
+      // SECURITY: Check if login request has expired
+      const now = Date.now();
+      if (loginRequest.expiresAt && now > loginRequest.expiresAt) {
+        throw new Error('Login request expired');
+      }
+
+      // Use the redirect URI that was stored with the login request
+      const redirectUri = loginRequest.redirectUri;
+      if (!redirectUri) {
+        throw new Error('No redirect URI found in login request');
+      }
+
+      // Exchange code for Google profile
+      const { profile, success } = await ctx.runAction(api.googleAuth.exchangeGoogleCode, {
+        code,
+        state,
+        redirectUri,
+      });
+      if (!success) throw new Error('Google OAuth failed');
+
+      // Find or create user and update session - using mutation
+      const loginResult = await ctx.runMutation(api.googleAuth.loginWithGoogle, {
+        profile,
+        sessionId: loginRequest.sessionId as SessionId, // SessionId type casting
+      });
+      if (!loginResult.success) throw new Error('Login failed');
+
+      // Mark login request as completed
+      await ctx.runMutation(api.auth.google.completeLoginRequest, {
+        loginRequestId: state as Id<'auth_loginRequests'>,
+        status: 'completed',
+      });
+
+      return {
+        success: true,
+        message: 'Login successful. You may close this window.',
+      };
+    } catch (err) {
+      // Mark login request as failed
+      await ctx.runMutation(api.auth.google.completeLoginRequest, {
+        loginRequestId: state as Id<'auth_loginRequests'>,
+        status: 'failed',
+        error: err instanceof Error ? err.message : 'Unknown error',
+      });
+
+      return {
+        success: false,
+        error: err instanceof Error ? err.message : 'Unknown error',
+      };
+    }
+  },
+});
+
+/**
+ * Handles Google OAuth callback for profile connect flow.
+ * Processes the OAuth code, exchanges it for a profile, connects the account to existing user, and marks the request as completed.
+ */
+export const handleGoogleConnectCallback = action({
+  args: {
+    code: v.string(),
+    state: v.string(), // This is the loginRequestId
+  },
+  handler: async (ctx, args) => {
+    const { code, state } = args;
+
+    try {
+      // Get the login request to extract sessionId and redirectUri
+      const loginRequest = await ctx.runQuery(api.auth.google.getLoginRequest, {
+        loginRequestId: state as Id<'auth_loginRequests'>,
+      });
+      if (!loginRequest || loginRequest.provider !== 'google') {
+        throw new Error('Invalid login request');
+      }
+
+      // SECURITY: Check if login request has expired
+      const now = Date.now();
+      if (loginRequest.expiresAt && now > loginRequest.expiresAt) {
+        throw new Error('Login request expired');
+      }
+
+      // Use the redirect URI that was stored with the login request
+      const redirectUri = loginRequest.redirectUri;
+      if (!redirectUri) {
+        throw new Error('No redirect URI found in login request');
+      }
+
+      // Exchange code for Google profile
+      const { profile, success } = await ctx.runAction(api.googleAuth.exchangeGoogleCode, {
+        code,
+        state,
+        redirectUri,
+      });
+      if (!success) throw new Error('Google OAuth failed');
+
+      // Connect Google account to existing user - using mutation
+      const connectResult = await ctx.runMutation(api.googleAuth.connectGoogle, {
+        profile,
+        sessionId: loginRequest.sessionId as SessionId, // SessionId type casting
+      });
+      if (!connectResult.success) throw new Error('Connect failed');
+
+      // Mark login request as completed
+      await ctx.runMutation(api.auth.google.completeLoginRequest, {
+        loginRequestId: state as Id<'auth_loginRequests'>,
+        status: 'completed',
+      });
+
+      return {
+        success: true,
+        message: 'Account connected successfully. You may close this window.',
+      };
+    } catch (err) {
+      // Mark login request as failed
+      await ctx.runMutation(api.auth.google.completeLoginRequest, {
+        loginRequestId: state as Id<'auth_loginRequests'>,
+        status: 'failed',
+        error: err instanceof Error ? err.message : 'Unknown error',
+      });
+
+      return {
+        success: false,
+        error: err instanceof Error ? err.message : 'Unknown error',
+      };
+    }
   },
 });
